move interfaces outside packages
