+++ src/main/java/ttt/gameplay/GameLoop.java
@@ -1,8 +1,8 @@
 package ttt.gameplay;
 
 import ttt.consoleui.ConsoleBoard;
-import ttt.consoleui.Display;
-import ttt.consoleui.Input;
+import ttt.Display;
+import ttt.Input;
 import ttt.game.GameEngine;
 import ttt.game.Marks;
