+++ src/test/java/ttt/ComputerPlayerTest.java
@@ -2,7 +2,6 @@ package ttt;
 
 import org.junit.Before;
 import org.junit.Test;
-import ttt.consoleui.Input;
 import ttt.game.Board;
 import ttt.game.Marks;
 import ttt.players.ComputerPlayer;
