+++ src/test/java/ttt/DisplayFake.java
@@ -1,6 +1,5 @@
 package ttt;
 
-import ttt.consoleui.Display;
 import ttt.game.Marks;
 
 public class DisplayFake implements Display {
