+++ src/test/java/ttt/GameEngineTest.java
@@ -5,7 +5,6 @@ import org.junit.Test;
 import ttt.game.Board;
 import ttt.game.GameEngine;
 import ttt.game.Marks;
-import ttt.game.Player;
 import ttt.players.HumanPlayer;
 
 import java.util.Arrays;
