+++ src/test/java/ttt/GameLoopTest.java
@@ -6,7 +6,6 @@ import ttt.consoleui.ConsoleBoard;
 import ttt.game.Board;
 import ttt.game.GameEngine;
 import ttt.game.Marks;
-import ttt.game.Player;
 import ttt.gameplay.GameLoop;
 import ttt.players.HumanPlayer;
