+++ src/test/java/ttt/HumanPlayerTest.java
@@ -3,7 +3,6 @@ package ttt;
 import org.junit.Test;
 import ttt.game.Board;
 import ttt.game.Marks;
-import ttt.game.Player;
 import ttt.players.HumanPlayer;
 
 import static org.junit.Assert.assertEquals;
