+++ src/test/java/ttt/InputFake.java
@@ -1,7 +1,5 @@
 package ttt;
 
-import ttt.consoleui.Input;
-
 import java.util.ArrayList;
 
 public class InputFake implements Input {
