+++ src/test/java/ttt/PlayerFactoryTest.java
@@ -2,7 +2,6 @@ package ttt;
 
 import org.junit.Test;
 import ttt.game.Marks;
-import ttt.game.Player;
 import ttt.players.*;
 
 import static junit.framework.TestCase.assertTrue;
